:art: Improving architecture
